ibmetrics/reader: handle "None" in date parser
The pandas date parser will give up on a column if it finds a string it
cannot parse and keep the column as a string (object).  The tsv data
for the subscriptions is read as strings so "None" interrupts the date
parser.

In our custom date reader, convert "None" to None so that the numpy date
conversion succeeds and we can have columns with datetime dtype and null
values.